Moved types into it's own directory.
